worker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
langchain==0.2.1
sentence-transformers==2.7.0
chromadb==0.5.23
pymongo==4.8.0